Move types to validation and types refact
