+++ src/controllers/presentation.ts
@@ -4,7 +4,7 @@ import { StatusCodes } from 'http-status-codes';
 import { check, validationResult, query } from './validator/index.js';
 import { IdentityServiceStrategySetup } from '../services/identity/index.js';
 import { CheqdW3CVerifiablePresentation } from '../services/w3c-presentation.js';
-import type { VerifyPresentationResponseBody } from '../types/shared.js';
+import type { VerifyPresentationResponseBody } from '../types/presentation.js';
 import { isIssuerDidDeactivated } from '../services/helpers.js';
 
 export class PresentationController {
