+++ src/types/credential.ts
@@ -1,0 +1,24 @@
+import type { VerificationPolicies } from '@veramo/core';
+import type { StatusOptions } from './credential-status.js';
+
+
+export interface CredentialRequest {
+	subjectDid: string;
+	attributes: Record<string, unknown>;
+	'@context'?: string[];
+	type?: string[];
+	expirationDate?: DateType;
+	issuerDid: string;
+	format: 'jsonld' | 'jwt';
+	credentialStatus?: StatusOptions;
+	credentialSchema?: string;
+	credentialName?: string;
+	credentialSummary?: string;
+}export interface VerificationOptions {
+	fetchRemoteContexts?: boolean;
+	policies?: VerificationPolicies;
+	domain?: string;
+	verifyStatus?: boolean;
+}
+export type DateType = string | Date;
+
