+++ src/types/presentation.ts
@@ -1,0 +1,5 @@
+
+export type VerifyPresentationResponseBody = {
+	verified: false;
+	error: string;
+};
